requirements-test.txt
# Requirements for running tests
pytest==7.3.1
pytest-flask==1.2.0
pytest-cov==4.1.0
coverage==7.2.7
termcolor==2.3.0
requests==2.31.0 